+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 0.26.2 - Unreleased
+### Features
+- Live updating of core options
+- Mounts `/libretro/core` as the directory where the core was loaded from
+- Adds `love.filesystem.getSaveDirectory()`
+
 ### Fixes
 - Fixed loading ChaiLove without active content